Tweaked btree traversal to visit each leaf at most once
We are already paying the memory cost of a fetched lfsr_rbyd_t
during btree so we can traverse inner btree nodes. But we are currently
just wasting this memory when we traverse leaf entries.

Instead, we can use this memory to cache the btree's current leaf node,
avoiding a btree walk until we've iterated over all rids in the current
leaf.

Think about this for a second:

1. We cache the root rbyd in lfsr_btree_t because we share the memory with
   a union and we always need to read the root during lookups.

2. We cache the leaf rbyds in lfsr_btree_traversal_t because we need the
   memory for traversing inner btree nodes.

The only btree nodes we don't cache during traversal are inner nodes
when the height of the btree >= 3.

If you're familiar with how btrees behave on storage, you know the
height because _exponentially_ less likely to grow as the tree gets
larger. It's entirely possible for btree traversal to simply never
traverse a non-cached btree node once your block size gets large
enough.

---

There may be a way to instead reclaim this memory, such as sharing this
memory with mtree traversal and higher layers, but the current API design
and hierarchy of C structs make this difficult. Maybe this is worth looking
into in the future, but at most it would save one lfsr_rbyd_t.

This also reverts some rid changes in btree lookups to share less code
but be a bit easier to reason about.